Minor grammatical change
I do note the difference in the following text
- Items in your Qobuz library will be marked as "in library" in Music Assistant.
- Marking an item as "in library" from the Music Assistant interface will also mark it as a favorite in Qobuz

I am not sure whether things are added to library or favourited? In the Deezer description it states favourites are synced back and forth and doesn’t make mention of the library. I don’t use either of these but it would useful to clarify exactly what happens.